Extracted request into a separate file
