+++ codeclimate/data_source_repository.go
@@ -1,11 +1,7 @@
 package codeclimate
 
 import (
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
 	"log"
-	"net/http"
 
 	"github.com/hashicorp/terraform/helper/schema"
 )
@@ -13,15 +9,6 @@ import (
 const (
 	codeClimateApiHost = "https://api.codeclimate.com/v1"
 )
-
-type ReadRepositoryResponse struct {
-	Data struct {
-		ID         string `json:"id"`
-		Attributes struct {
-			TestReporterID string `json:"test_reporter_id"`
-		} `json:"attributes"`
-	} `json:"data"`
-}
 
 func dataSourceRepository() *schema.Resource {
 	return &schema.Resource{
@@ -41,8 +28,8 @@ func dataSourceRepository() *schema.Resource {
 }
 
 func dataSourceRepositoryRead(d *schema.ResourceData, m interface{}) error {
-	repoId := d.Get("repository_id").(string)
-	repositoryData, err := getRepository(m.(Config).apiKey, repoId)
+	repositoryId := d.Get("repository_id").(string)
+	repositoryData, err := getRepository(m.(Config).apiKey, repositoryId)
 	if err != nil {
 		log.Println(err)
 		return err
@@ -53,44 +40,4 @@ func dataSourceRepositoryRead(d *schema.ResourceData, m interface{}) error {
 	d.Set("test_reporter_id", repositoryData.(ReadRepositoryResponse).Data.Attributes.TestReporterID)
 
 	return err
-}
-
-func getRepository(apiKey string, repoId string) (interface{}, error) {
-	var repositoryData ReadRepositoryResponse
-
-	// TODO: Extract into a client
-	client := &http.Client{}
-
-	req, err := http.NewRequest("GET", fmt.Sprintf("%s/repos/%s", codeClimateApiHost, repoId), nil)
-
-	if err != nil {
-		log.Println(err)
-		return nil, err
-	}
-
-	req.Header.Add("Accept", `W/"application/vnd.api+json"`)
-	req.Header.Add("Authorization", fmt.Sprintf("Token token=%s", apiKey))
-
-	resp, err := client.Do(req)
-
-	if err != nil {
-		log.Println(err)
-		return nil, err
-	}
-
-	defer resp.Body.Close()
-	data, err := ioutil.ReadAll(resp.Body)
-
-	if err != nil {
-		log.Println(err)
-		return nil, err
-	}
-
-	err = json.Unmarshal(data, &repositoryData)
-	if err != nil {
-		log.Println(err)
-		return nil, err
-	}
-
-	return repositoryData, nil
 }
